restructurate view builder logic
