+++ src/jca/springframework/builder/view/JsonViewBuilder.java
@@ -1,0 +1,29 @@
+package jca.springframework.builder.view;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
+import jca.springframework.mapping.MappingAnnotation;
+import jca.springframework.view.ModelAndView;
+import jca.springframework.view.RestApiView;
+import jca.springframework.view.View;
+
+public class JsonViewBuilder extends ViewBuilder {
+    final static Gson gson = new GsonBuilder().create();
+    public static Gson getGson() {
+        return gson;
+    }
+    @Override
+    public View buildView(Object obj, MappingAnnotation mappingAnnotation) {
+        View view= null;
+        if (obj instanceof ModelAndView){
+            ModelAndView modelAndView = (ModelAndView) obj;
+            view = new RestApiView( getGson().toJson(modelAndView.getData()));
+        }
+        else {
+            view = new RestApiView( getGson().toJson(obj) );            
+        }
+        return view;
+    }
+    
+}
